requirements.txt
crawl4ai
feedparser
google-generativeai
libsql-client
python-dotenv
asyncio
requests